+++ magik-language-server/src/main/java/nl/ramsolutions/sw/magik/languageserver/inlayhint/ArgumentNameInlayHintSupplier.java
@@ -4,6 +4,7 @@ import com.sonar.sslr.api.AstNode;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
+import java.util.stream.Collectors;
 import java.util.stream.Stream;
 import nl.ramsolutions.sw.MagikToolsProperties;
 import nl.ramsolutions.sw.magik.MagikTypedFile;
@@ -61,8 +62,13 @@ class ArgumentNameInlayHintSupplier {
     final LocalTypeReasonerState reasonerState = magikFile.getTypeReasonerState();
     final AstNode previousSiblingNode = methodInvocationNode.getPreviousSibling();
     final ExpressionResultString result = reasonerState.getNodeType(previousSiblingNode);
-    final TypeString typeStr = result.get(0, TypeString.UNDEFINED);
-    if (typeStr == TypeString.UNDEFINED) {
+    final TypeString typeString = result.get(0, TypeString.UNDEFINED);
+    if (typeString == TypeString.UNDEFINED) {
+      return Stream.of();
+    }
+
+    // Only supply for single TypeStrings.
+    if (!typeString.isSingle()) {
       return Stream.of();
     }
 
@@ -70,8 +76,10 @@ class ArgumentNameInlayHintSupplier {
     final MethodInvocationNodeHelper helper = new MethodInvocationNodeHelper(methodInvocationNode);
     final String methodName = helper.getMethodName();
     final TypeStringResolver resolver = magikFile.getTypeStringResolver();
-    final Collection<MethodDefinition> methodDefinitions =
-        resolver.getMethodDefinitions(typeStr, methodName);
+    final Collection<MethodDefinition> methodDefinitions = typeString.getCombinedTypes().stream()
+        .map(typeStr -> resolver.getRespondingMethodDefinitions(typeStr, methodName))
+        .flatMap(Collection::stream)
+        .collect(Collectors.toSet());
     final MethodDefinition methodDefinition = methodDefinitions.stream().findAny().orElse(null);
     if (methodDefinition == null) {
       return Stream.of();