+++ magik-squid/src/main/java/nl/ramsolutions/sw/magik/analysis/typing/reasoner/restrictions/InverseScopeEntryTypeRestriction.java
@@ -30,7 +30,7 @@ public class InverseScopeEntryTypeRestriction implements TypeRestriction {
   /**
    * Get the inverse of ourselves.
    *
-   * @return
+   * @return The inverse of ourselves.
    */
   @Override
   public TypeRestriction not() {